pyproject.toml
[project]
name = "timetable"
version = "0.1.0"
description = "MSU timetable scraper with per-group ICS server and Telegram bot"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "icalendar>=5.0.0",
    "python-dotenv>=1.0.1",
    "selenium>=4.21.0",
    "loguru>=0.7.2",
    "SQLAlchemy>=2.0.0",
    "pg8000>=1.30.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.5.0",
    "black>=24.4.0",
    "isort>=5.13.0",
    "pytest>=8.0.0",
]

[project.scripts]
timetable = "main:run"

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.ruff]
line-length = 100
target-version = "py313"
fix = true

[tool.ruff.lint]
select = [
    "E", "F", "I",  # pycodestyle, pyflakes, isort
    "B",              # flake8-bugbear
    "UP",             # pyupgrade
]
ignore = [
    "E501",  # line length handled by formatter
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true
py-modules = ["main"]

[tool.setuptools.packages.find]
include = [
    "parse*",
    "serve*",
    "sync*",
    "utils*",
]

[tool.isort]
profile = "black"
line_length = 100
